dev-requirements.txt
ruff
isort
django-dynamic-fixture
coverage